+++ analyzers/tests/SonarAnalyzer.UnitTest/TestCases/SymbolicExecution/Roslyn/PublicMethodArgumentsShouldBeCheckedForNull.cs
@@ -1,609 +1,389 @@
 ﻿using System;
+using System.Text;
 
-namespace Tests.Diagnostics
+public class Program
 {
-    public class Program
+    private object field = null;
+    private static object staticField = null;
+
+    public void NotCompliantCases(object o, Exception e)
     {
-        private object field = null;
-        private static object staticField = null;
+        o.ToString(); // FIXME Non-compliant {{Refactor this method to add validation of parameter 'o' before using it.}}
 
-        public void NotCompliantCases(object o, Exception e)
+        Bar(o); // Compliant, we care about dereference only
+
+        throw e; // FIXME Non-compliant
+    }
+
+    public void Bar(object o) { }
+
+    protected void NotCompliantCases_Nonpublic(object o)
+    {
+        o.ToString(); // FIXME Non-compliant
+    }
+
+    private void CompliantCases_Private(object o)
+    {
+        o.ToString(); // Compliant, not public
+    }
+
+    protected internal void NotCompliantCases_ProtectedInternal(object o)
+    {
+        o.ToString(); // FIXME Non-compliant
+    }
+
+    internal void CompliantCases_Internal(object o)
+    {
+        o.ToString(); // Compliant, not public
+    }
+
+    public void CompliantCases(bool b, object o1, object o2, object o3, object o4, Exception e)
+    {
+        if (o1 != null)
         {
-            o.ToString(); // FIXME Non-compliant {{Refactor this method to add validation of parameter 'o' before using it.}}
-
-            Bar(o); // Compliant, we care about dereference only
-
-            throw e; // FIXME Non-compliant
+            o1.ToString(); // Compliant, we did the check
         }
 
-        public void Bar(object o) { }
+        o2 = o2 ?? new object();
+        o2.ToString(); // Compliant, we coalesce
 
-        protected void NotCompliantCases_Nonpublic(object o)
+        if (o3 == null)
         {
-            o.ToString(); // FIXME Non-compliant
+            throw new Exception();
         }
 
-        private void CompliantCases_Private(object o)
+        o3.ToString(); // Compliant, we did the check
+
+        if (e != null)
         {
-            o.ToString(); // Compliant, not public
+            throw e; // Compliant
         }
 
-        protected internal void NotCompliantCases_ProtectedInternal(object o)
+        o4?.ToString(); // Compliant, conditional operator
+
+        b.ToString(); // Compliant, bool cannot be null
+
+        object v = null;
+        v.ToString(); // Compliant, we don't care about local variables
+
+        field.ToString(); // Compliant
+
+        Program.staticField.ToString(); // Compliant
+    }
+
+    public void MoreCompliantCases(string s1, string s2)
+    {
+        if (string.IsNullOrEmpty(s1))
         {
-            o.ToString(); // FIXME Non-compliant
+            s1.ToString(); // FIXME Non-compliant, could be null
+        }
+        else
+        {
+            s1.ToString(); // Compliant
         }
 
-        internal void CompliantCases_Internal(object o)
+        if (string.IsNullOrWhiteSpace(s2))
         {
-            o.ToString(); // Compliant, not public
+            s2.ToString(); // FIXME Non-compliant, could be null
         }
-
-        public void CompliantCases(bool b, object o1, object o2, object o3, object o4, Exception e)
+        else
         {
-            if (o1 != null)
-            {
-                o1.ToString(); // Compliant, we did the check
-            }
-
-            o2 = o2 ?? new object();
-            o2.ToString(); // Compliant, we coalesce
-
-            if (o3 == null)
-            {
-                throw new Exception();
-            }
-
-            o3.ToString(); // Compliant, we did the check
-
-            if (e != null)
-            {
-                throw e; // Compliant
-            }
-
-            o4?.ToString(); // Compliant, conditional operator
-
-            b.ToString(); // Compliant, bool cannot be null
-
-            object v = null;
-            v.ToString(); // Compliant, we don't care about local variables
-
-            field.ToString(); // Compliant
-
-            Program.staticField.ToString(); // Compliant
-        }
-
-        public void MoreCompliantCases(string s1, string s2)
-        {
-            if (string.IsNullOrEmpty(s1))
-            {
-                s1.ToString(); // FIXME Non-compliant, could be null
-            }
-            else
-            {
-                s1.ToString(); // Compliant
-            }
-
-            if (string.IsNullOrWhiteSpace(s2))
-            {
-                s2.ToString(); // FIXME Non-compliant, could be null
-            }
-            else
-            {
-                s2.ToString(); // Compliant
-            }
-        }
-
-        public Program(int i) { }
-
-        public Program(string s) : this(s.Length) { }   // FIXME Non-compliant {{Refactor this constructor to avoid using members of parameter 's' because it could be null.}}
-
-        public void NonCompliant1(object o)
-        {
-            var c = o?.ToString()?.IsNormalized();
-            if (c == null)
-            {
-                o.GetType().GetMethods();  // FIXME Non-compliant
-            }
-        }
-
-        public void Compliant1(object o)
-        {
-            var c = o?.ToString()?.IsNormalized();
-            if (c != null)
-            {
-                o.GetType().GetMethods();
-            }
+            s2.ToString(); // Compliant
         }
     }
 
-    [AttributeUsage(AttributeTargets.Parameter)]
-    public sealed class ValidatedNotNullAttribute : Attribute { }
+    public Program(int i) { }
 
-    [AttributeUsage(AttributeTargets.Parameter)]
-    public sealed class FooAttribute : Attribute { }
+    public Program(string s) : this(s.Length) { }   // FIXME Non-compliant {{Refactor this constructor to avoid using members of parameter 's' because it could be null.}}
 
-    [AttributeUsage(AttributeTargets.Parameter)]
-    public sealed class BarAttribute : Attribute { }
-
-    public static class GuardExtensionClass
+    public void NonCompliant1(object o)
     {
-        public static void GuardExtension([ValidatedNotNull] this string value) { }
-        public static void GuardExtensionMoreAttributes([Foo][ValidatedNotNull][Bar] this string value) { }
-        public static void GuardExtensionMoreArguments([ValidatedNotNull] this string value, string foo, string bar) { }
-    }
-
-    public class GuardedTests
-    {
-        public void Guarded(string s1, string s2, string s3, string s4, string s5, string s6, string s7)
+        var c = o?.ToString()?.IsNormalized();
+        if (c == null)
         {
-            Guard1(s1);
-            s1.ToUpper();
-
-            Guard2(s2, "s2");
-            s2.ToUpper();
-
-            Guard3("s3", s3);
-            s3.ToUpper();
-
-            Guard4(s4);
-            s4.ToUpper();
-
-            s5.GuardExtension();
-            s5.ToUpper();
-
-            s6.GuardExtensionMoreAttributes();
-            s6.ToUpper();
-
-            s7.GuardExtensionMoreArguments(null, null);
-            s7.ToUpper();
-        }
-
-        public void Guard1<T>([ValidatedNotNull] T value) where T : class { }
-
-        public void Guard2<T>([ValidatedNotNull] T value, string name) where T : class { }
-
-        public void Guard3<T>(string name, [ValidatedNotNull] T value) where T : class { }
-
-        public static void Guard4<T>([ValidatedNotNull] T value) where T : class { }
-    }
-
-    public class ReproIssue2476
-    {
-        public void Foo(params string[] infixes)
-        {
-            if (infixes == null || infixes.Length == 0)
-            {
-                infixes = new string[1];
-            }
-            else
-            {
-                Array.Resize(ref infixes, infixes.Length + 1);
-            }
-            // more stuff
-        }
-
-        public void Method(ref string s, int x) { }
-        public void Method1(string infixes)
-        {
-            if (infixes != null)
-            {
-                Method(ref infixes, infixes.Length);
-                var x = infixes.Length; // FIXME Non-compliant when passed by ref can be set to null
-            }
-
-        }
-
-        public void Method2(string infixes)
-        {
-            if (infixes == null)
-            {
-                Method(ref infixes, infixes.Length); // FIXME Non-compliant
-                var x = infixes.Length;
-            }
-        }
-
-        public void Method3(string infixes)
-        {
-            Method(ref infixes, infixes.Length); // FIXME Non-compliant
-            var x = infixes.Length; // FIXME Non-compliant
-        }
-
-        public void Method4(string contentType)
-        {
-            if (string.IsNullOrEmpty(contentType))
-            {
-                throw new ArgumentException("inputString cannot be null or empty", contentType);
-            }
-
-            if (contentType.Equals("*"))
-            {
-                contentType = "*/*";
-            }
-
-            var parts = contentType.Split('/', ';');
-        }
-
-        public void Method5(string contentType)
-        {
-            if (string.IsNullOrEmpty(contentType))
-            {
-                throw new ArgumentException("inputString cannot be null or empty", contentType);
-            }
-
-            if (contentType.Equals("*"))
-            {
-                contentType = "";
-            }
-
-            var parts = contentType.Split('/', ';');
-        }
-
-        public void Method6(string contentType)
-        {
-            if (string.IsNullOrEmpty(contentType))
-            {
-                throw new ArgumentException("inputString cannot be null or empty", contentType);
-            }
-
-            if (contentType.Equals("*"))
-            {
-                contentType = null;
-            }
-
-            var parts = contentType.Split('/', ';'); // FIXME Non-compliant
+            o.GetType().GetMethods();  // FIXME Non-compliant
         }
     }
 
-    // https://github.com/SonarSource/sonar-dotnet/issues/2591
-    public class ReproIssue2591
+    public void Compliant1(object o)
     {
-        public string FooWithStringTrim(string name)
+        var c = o?.ToString()?.IsNormalized();
+        if (c != null)
         {
-            if (name == null)
-            {
-                name = Guid.NewGuid().ToString("N");
-            }
-
-            return name.Trim(); // FIXME Non-compliant FP
-        }
-
-        public string FooWithStringJoin(string name)
-        {
-            if (name == null)
-            {
-                name = Guid.NewGuid().ToString("N");
-            }
-
-            return string.Join("_", name.Split(System.IO.Path.GetInvalidFileNameChars())); // FIXME Non-compliant FP
-        }
-
-        public string FooWithObject(object name)
-        {
-            if (name == null)
-            {
-                name = Guid.NewGuid().ToString("N");
-            }
-
-            return name.ToString(); // FIXME Non-compliant FP
-        }
-    }
-
-    // https://github.com/SonarSource/sonar-dotnet/issues/2670
-    public class ReproIssue2670
-    {
-        public static void BooleanEqualityComparisonIsNullOrEmpty(string argument, bool b)
-        {
-            if (string.IsNullOrEmpty(argument) == true)
-            {
-                return;
-            }
-            if (b)
-            {
-                int index = argument.LastIndexOf('c'); // FIXME Non-compliant FP
-            }
-        }
-
-        public static void BooleanEqualityComparisonWithIsNullOrWhiteSpace(string argument, bool b)
-        {
-            if (string.IsNullOrWhiteSpace(argument) == true)
-            {
-                return;
-            }
-            if (b)
-            {
-                int index = argument.LastIndexOf('c'); // FIXME Non-compliant FP
-            }
-        }
-
-        public static void NoComparison(string argument, bool b)
-        {
-            if (string.IsNullOrEmpty(argument))
-            {
-                return;
-            }
-            if (b)
-            {
-                int index = argument.LastIndexOf('c');
-            }
-        }
-    }
-
-    // https://github.com/SonarSource/sonar-dotnet/issues/2775
-    public class ReproWithIs
-    {
-        public override bool Equals(object obj)
-        {
-            var equals = (obj is string) && (obj.GetHashCode() == GetHashCode()); // FIXME Non-compliant FP
-            if (equals)
-            {
-                // do stuff
-            }
-            return equals;
-        }
-    }
-
-    public class ReproWithIsNullOrEmpty
-    {
-        public static string SanitiseIMDbId(string imdbId)
-        {
-            if (string.IsNullOrEmpty(imdbId))
-                return string.Empty;
-
-            if (imdbId.StartsWith(" "))
-                imdbId = string.Concat("tt", imdbId);
-
-            if (imdbId.Length != 9) // FIXME Non-compliant - FP
-                imdbId = string.Empty;
-
-            return imdbId;
+            o.GetType().GetMethods();
         }
     }
 }
 
-namespace CSharp8
+[AttributeUsage(AttributeTargets.Parameter)]
+public sealed class ValidatedNotNullAttribute : Attribute { }
+
+[AttributeUsage(AttributeTargets.Parameter)]
+public sealed class FooAttribute : Attribute { }
+
+[AttributeUsage(AttributeTargets.Parameter)]
+public sealed class BarAttribute : Attribute { }
+
+public static class GuardExtensionClass
 {
-    public class NullCoalescenceAssignment
+    public static void GuardExtension([ValidatedNotNull] this string value) { }
+    public static void GuardExtensionMoreAttributes([Foo][ValidatedNotNull][Bar] this string value) { }
+    public static void GuardExtensionMoreArguments([ValidatedNotNull] this string value, string foo, string bar) { }
+}
+
+public class GuardedTests
+{
+    public void Guarded(string s1, string s2, string s3, string s4, string s5, string s6, string s7)
     {
-        public void NullCoalescenceAssignment_NotNull(string s1)
+        Guard1(s1);
+        s1.ToUpper();
+
+        Guard2(s2, "s2");
+        s2.ToUpper();
+
+        Guard3("s3", s3);
+        s3.ToUpper();
+
+        Guard4(s4);
+        s4.ToUpper();
+
+        s5.GuardExtension();
+        s5.ToUpper();
+
+        s6.GuardExtensionMoreAttributes();
+        s6.ToUpper();
+
+        s7.GuardExtensionMoreArguments(null, null);
+        s7.ToUpper();
+    }
+
+    public void Guard1<T>([ValidatedNotNull] T value) where T : class { }
+
+    public void Guard2<T>([ValidatedNotNull] T value, string name) where T : class { }
+
+    public void Guard3<T>(string name, [ValidatedNotNull] T value) where T : class { }
+
+    public static void Guard4<T>([ValidatedNotNull] T value) where T : class { }
+}
+
+public class ReproIssue2476
+{
+    public void Foo(params string[] infixes)
+    {
+        if (infixes == null || infixes.Length == 0)
         {
-            s1 ??= "N/A";
-            s1.ToString(); // Compliant
+            infixes = new string[1];
+        }
+        else
+        {
+            Array.Resize(ref infixes, infixes.Length + 1);
+        }
+        // more stuff
+    }
+
+    public void Method(ref string s, int x) { }
+    public void Method1(string infixes)
+    {
+        if (infixes != null)
+        {
+            Method(ref infixes, infixes.Length);
+            var x = infixes.Length; // FIXME Non-compliant when passed by ref can be set to null
         }
 
-        public void NullCoalescenceAssignment_Null(string s1)
-        {
-            s1 ??= null;
-            s1.ToString(); // FIXME Non-compliant
-        }
+    }
 
-        public void InsideIf(string str)
+    public void Method2(string infixes)
+    {
+        if (infixes == null)
         {
-            if (str != null)
-            {
-                str ??= null;
-                str.ToString(); // Compliant, we know str is not null
-            }
-
-            if (str == null)
-            {
-                str ??= "foo";
-                str.ToString(); // Compliant, assigned foo
-            }
+            Method(ref infixes, infixes.Length); // FIXME Non-compliant
+            var x = infixes.Length;
         }
     }
 
-    public interface IWithDefaultMembers
+    public void Method3(string infixes)
     {
-        decimal Count { get; set; }
-        decimal Price { get; set; }
+        Method(ref infixes, infixes.Length); // FIXME Non-compliant
+        var x = infixes.Length; // FIXME Non-compliant
+    }
 
-        void Reset(string s)
+    public void Method4(string contentType)
+    {
+        if (string.IsNullOrEmpty(contentType))
         {
-            s.ToString(); // FIXME Non-compliant
+            throw new ArgumentException("inputString cannot be null or empty", contentType);
+        }
+
+        if (contentType.Equals("*"))
+        {
+            contentType = "*/*";
+        }
+
+        var parts = contentType.Split('/', ';');
+    }
+
+    public void Method5(string contentType)
+    {
+        if (string.IsNullOrEmpty(contentType))
+        {
+            throw new ArgumentException("inputString cannot be null or empty", contentType);
+        }
+
+        if (contentType.Equals("*"))
+        {
+            contentType = "";
+        }
+
+        var parts = contentType.Split('/', ';');
+    }
+
+    public void Method6(string contentType)
+    {
+        if (string.IsNullOrEmpty(contentType))
+        {
+            throw new ArgumentException("inputString cannot be null or empty", contentType);
+        }
+
+        if (contentType.Equals("*"))
+        {
+            contentType = null;
+        }
+
+        var parts = contentType.Split('/', ';'); // FIXME Non-compliant
+    }
+}
+
+// https://github.com/SonarSource/sonar-dotnet/issues/2591
+public class ReproIssue2591
+{
+    public string FooWithStringTrim(string name)
+    {
+        if (name == null)
+        {
+            name = Guid.NewGuid().ToString("N");
+        }
+
+        return name.Trim(); // FIXME Non-compliant FP
+    }
+
+    public string FooWithStringJoin(string name)
+    {
+        if (name == null)
+        {
+            name = Guid.NewGuid().ToString("N");
+        }
+
+        return string.Join("_", name.Split(System.IO.Path.GetInvalidFileNameChars())); // FIXME Non-compliant FP
+    }
+
+    public string FooWithObject(object name)
+    {
+        if (name == null)
+        {
+            name = Guid.NewGuid().ToString("N");
+        }
+
+        return name.ToString(); // FIXME Non-compliant FP
+    }
+}
+
+// https://github.com/SonarSource/sonar-dotnet/issues/2670
+public class ReproIssue2670
+{
+    public static void BooleanEqualityComparisonIsNullOrEmpty(string argument, bool b)
+    {
+        if (string.IsNullOrEmpty(argument) == true)
+        {
+            return;
+        }
+        if (b)
+        {
+            int index = argument.LastIndexOf('c'); // FIXME Non-compliant FP
         }
     }
 
-    public class LocalStaticFunctions
+    public static void BooleanEqualityComparisonWithIsNullOrWhiteSpace(string argument, bool b)
     {
-        public void Method(object arg)
+        if (string.IsNullOrWhiteSpace(argument) == true)
         {
-            string LocalFunction(object o)
-            {
-                return o.ToString(); // Compliant - FN: local functions are not supported by the CFG
-            }
-
-            static string LocalStaticFunction(object o)
-            {
-                return o.ToString(); // Compliant - FN: local functions are not supported by the CFG
-            }
+            return;
+        }
+        if (b)
+        {
+            int index = argument.LastIndexOf('c'); // FIXME Non-compliant FP
         }
     }
 
-    public class Address
+    public static void NoComparison(string argument, bool b)
     {
-        public string Name { get; }
-
-        public string State { get; }
-
-        public void Deconstruct(out string name, out string state) =>
-            (name, state) = (Name, State);
-    }
-
-    public class Person
-    {
-        public string Name { get; }
-
-        public Address Address { get; }
-
-        public void Deconstruct(out string name, out Address address) =>
-            (name, address) = (Name, Address);
-    }
-
-    public class SwitchExpressions
-    {
-        public void OnlyDiscardBranch_Noncompliant(string s, bool b)
+        if (string.IsNullOrEmpty(argument))
         {
-            var result = b switch
-            {
-                _ => s.ToString() // FIXME Non-compliant
-            };
+            return;
         }
-
-        public void MultipleBranches_Noncompliant(string s, int val)
+        if (b)
         {
-            var result = val switch
-            {
-                1 => "a",
-                2 => s.ToString(), // FIXME Non-compliant
-                _ => "b"
-            };
-        }
-
-        public void Nested_Noncompliant(string s, int val, bool condition)
-        {
-            var result = val switch
-            {
-                1 => "a",
-                2 => condition switch
-                {
-                    _ => s.ToString() // FIXME Non-compliant
-                },
-                _ => "b"
-            };
-        }
-
-        public void MultipleBranches_HandleNull(string s, int val)
-        {
-            var result = s switch
-            {
-                null => s.ToString(), // FIXME Non-compliant
-                _ => s.ToString() // Compliant as null was already handled
-            };
-        }
-
-        public void MultipleBranches_Compliant(string s, int val)
-        {
-            var result = val switch
-            {
-                1 => "a",
-                2 => s == null ? string.Empty : s.ToString(),
-                _ => "b"
-            };
-        }
-
-        public string MultipleBranches_PropertyPattern(Address address, string s)
-        {
-            return address switch
-            {
-                { State: "WA" } addr => s.ToString(), // FIXME Non-compliant
-                _ => string.Empty
-            };
-        }
-
-        public string MultipleBranches_PropertyPattern_FP(string s)
-        {
-            return s switch
-            {
-                { Length: 5 } => s.ToString(), // FIXME Non-compliant - FP we know that the length is 5 so the string cannot be null
-                _ => string.Empty
-            };
-        }
-
-        public string MultipleBranches_RecursivePattern(Person person, string s)
-        {
-            return person switch
-            {
-                { Address: { State: "WA" } } pers => s.ToString(), // FIXME Non-compliant
-                _ => string.Empty
-            };
-        }
-
-        public string MultipleBranches_TuplePattern(Address address, string s)
-        {
-            return address switch
-            {
-                var (name, state) => s.ToString(), // Compliant - FN
-                _ => string.Empty
-            };
-        }
-
-        public string MultipleBranches_WhenClause(Address address, string s)
-        {
-            return address switch
-            {
-                Address addr when addr.Name.Length > 0 => s.ToString(), // FIXME Non-compliant
-                Address addr when addr.Name.Length == s.Length => string.Empty, // FIXME Non-compliant
-                _ => string.Empty
-            };
-        }
-
-        public string MultipleBranches_VarDeclaration(Address address, string s)
-        {
-            return address switch
-            {
-                Address addr => s.ToString(), // FIXME Non-compliant
-                _ => string.Empty
-            };
-        }
-
-        public string TwoBranches_NoDefault(bool condition, string s)
-        {
-            return condition switch
-            {
-                true => s.ToString(), // FIXME Non-compliant
-                false => s.ToString() // FIXME Non-compliant
-            };
+            int index = argument.LastIndexOf('c');
         }
     }
+}
 
-    public class SwitchStatement
+// https://github.com/SonarSource/sonar-dotnet/issues/2775
+public class ReproWithIs
+{
+    public override bool Equals(object obj)
     {
-        public void Test(string s)
+        var equals = (obj is string) && (obj.GetHashCode() == GetHashCode()); // FIXME Non-compliant FP
+        if (equals)
         {
-            switch (s)
-            {
-                case null:
-                    break;
-
-                default:
-                    s.ToString(); // Compliant - the null is handled by the case null branch.
-                    break;
-            }
+            // do stuff
         }
+        return equals;
+    }
+}
+
+public class ReproWithIsNullOrEmpty
+{
+    public static string SanitiseIMDbId(string imdbId)
+    {
+        if (string.IsNullOrEmpty(imdbId))
+            return string.Empty;
+
+        if (imdbId.StartsWith(" "))
+            imdbId = string.Concat("tt", imdbId);
+
+        if (imdbId.Length != 9) // FIXME Non-compliant - FP
+            imdbId = string.Empty;
+
+        return imdbId;
     }
 }
 
 // https://github.com/SonarSource/sonar-dotnet/issues/3400
-namespace Repro_3400
+public class Repro_3400
 {
-    using System.Text;
-
-    public class Repro_3400
+    public void ReassignedFromMethod(StringBuilder parameter)
     {
-        public void ReassignedFromMethod(StringBuilder parameter)
-        {
-            parameter = Create();
-            parameter.Capacity = 1; // FIXME Non-compliant FP
-        }
-
-        public void ReassignedFromConstructor(StringBuilder parameter)
-        {
-            parameter = new StringBuilder();
-            parameter.Capacity = 1;
-        }
-
-        public void ReassignedFromMethodOut(out StringBuilder parameter)
-        {
-            parameter = Create();
-            parameter.Capacity = 1; // FIXME Non-compliant FP
-        }
-
-        public void ReassignedFromConstructorOut(out StringBuilder parameter)
-        {
-            parameter = new StringBuilder();
-            parameter.Capacity = 1;
-        }
-
-        private StringBuilder Create() => null;
+        parameter = Create();
+        parameter.Capacity = 1; // FIXME Non-compliant FP
     }
+
+    public void ReassignedFromConstructor(StringBuilder parameter)
+    {
+        parameter = new StringBuilder();
+        parameter.Capacity = 1;
+    }
+
+    public void ReassignedFromMethodOut(out StringBuilder parameter)
+    {
+        parameter = Create();
+        parameter.Capacity = 1; // FIXME Non-compliant FP
+    }
+
+    public void ReassignedFromConstructorOut(out StringBuilder parameter)
+    {
+        parameter = new StringBuilder();
+        parameter.Capacity = 1;
+    }
+
+    private StringBuilder Create() => null;
 }
